remove dropDown.jsx
